+++ src/main/java/com/atlauncher/gui/tabs/settings/GeneralSettingsTab.java
@@ -63,6 +63,7 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
     private final JCheckBox disableCustomFonts;
     private final JCheckBox rememberWindowSizePosition;
     private final JCheckBox useNativeFilePicker;
+    private final JCheckBox enableAddedModsByDefault;
 
     public GeneralSettingsTab() {
         // Language
@@ -475,6 +476,24 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
         useNativeFilePicker = new JCheckBox();
         useNativeFilePicker.setSelected(App.settings.useNativeFilePicker);
         add(useNativeFilePicker, gbc);
+
+        // Enable added mods by default
+
+        gbc.gridx = 0;
+        gbc.gridy++;
+        gbc.insets = UIConstants.LABEL_INSETS;
+        gbc.anchor = GridBagConstraints.BASELINE_TRAILING;
+        JLabelWithHover enableAddedModsByDefaultLabel = new JLabelWithHover(GetText.tr("Enable Added Mods By Default?"),
+                HELP_ICON, new HTMLBuilder().center().split(100)
+                        .text(GetText.tr("When adding mods manually, should they be enabled automatically?")).build());
+        add(enableAddedModsByDefaultLabel, gbc);
+
+        gbc.gridx++;
+        gbc.insets = UIConstants.CHECKBOX_FIELD_INSETS;
+        gbc.anchor = GridBagConstraints.BASELINE_LEADING;
+        enableAddedModsByDefault = new JCheckBox();
+        enableAddedModsByDefault.setSelected(App.settings.enableAddedModsByDefault);
+        add(enableAddedModsByDefault, gbc);
     }
 
     @SuppressWarnings("unchecked")
@@ -536,6 +555,7 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
         }
 
         App.settings.useNativeFilePicker = useNativeFilePicker.isSelected();
+        App.settings.enableAddedModsByDefault = enableAddedModsByDefault.isSelected();
     }
 
     @Override